Added PRNG implementations to mach.range.random
Implementations include LCG, mersenne twister, and xorshift PRNGs

If a PRNG range is constructed without any explicit seeds then mach.range.seed is leveraged to generate a seed from the current time, the process id, and the current thread id. In 90% of cases this is exactly what you want to do. In 5% of cases you want the same sequence every time you initialize the PRNG in which case you'd specify the seed manually, and in 5% of cases you probably need to think about cryptographic security and should use another implementation altogether which puts hardware RNGs to use, or which uses platform-specific sources of entropy for its random number generation.

Support for the latter two may come later on.