Can now chain an iterable of iterables with mach.range.chain
In the process I ended up making chain its own package under mach.range

If you try to chain a single iterable of iterables behavior will default to chaining the iterables contained within, i.e. flatten but if you want it to be interpreted as a single element in a chain of ranges of disparate types, use chainrange explicitly instead of range. Though why you'd ever end up doing that is beyond me