Fixed loading textures from SDL_Surface data
It was working before on Win7, but color channels were screwed up in OSX. Now it works for both. The solution needs some refining, but I'll get to that in a bit.

Note to self - GL_UNSIGNED_INT_8_8_8_8 fixed things here, but for other pixel formats GL_UNSIGNED_INT_8_8_8_8_REV may be preferable or outright necessary