Cleanup, more test, and more docs in mach.range.pad
Also added a remaining property, which was heretofore notably missing

Left a TODO regarding a funny case that should probably produce an error, involving attempting to pad the front of a partially-consumed range

At some point I will probably want to make an entirely separate range for padding only the back. Might be wise to do the same for front, if only for optimization purposes.